+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @authhero/demo
 
+## 0.0.6
+
+### Patch Changes
+
+- authhero@0.2.8
+
 ## 0.0.5
 
 ### Patch Changes