+++ packages/authhero/CHANGELOG.md
@@ -1,5 +1,12 @@
 # authhero
 
+## 0.2.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.10.4
+
 ## 0.2.7
 
 ### Patch Changes